added sun with proper lighting and bloom
