+++ package.json
@@ -16,6 +16,7 @@
     "@mui/material": "^6.2.0",
     "@react-three/drei": "^9.120.4",
     "@react-three/fiber": "^8.17.10",
+    "@react-three/postprocessing": "^2.16.5",
     "@types/three": "^0.170.0",
     "leva": "^0.9.35",
     "node-fetch": "^3.3.2",
