+++ src/r3f/Satellite.tsx
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react';
 import * as THREE from 'three';
-import { useLoader, useFrame } from '@react-three/fiber';
+import { useLoader } from '@react-three/fiber';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader';
 
 interface SatelliteProps {
@@ -8,18 +8,12 @@ interface SatelliteProps {
   stopRotation?: boolean;
 }
 
-const Satellite: React.FC<SatelliteProps> = ({ rotationSpeed = 1, stopRotation = false }) => {
+const Satellite: React.FC<SatelliteProps> = () => {
   const gltf = useLoader(GLTFLoader, '/models/ISS_stationary.glb');
-  const scaleSize = [0.001, 0.001, 0.001];
+  const scaleSize = [0.0005, 0.0005, 0.0005];
   const mesh = useRef<THREE.Mesh>(null);
 
-  useFrame(() => {
-    if (mesh.current && !stopRotation) {
-      mesh.current.rotation.y += 0.0005 * rotationSpeed;
-    }
-  });
-
-  gltf.scene.position.x = 3;
+  gltf.scene.position.x = 1.6;
 
   return <primitive object={gltf.scene} scale={scaleSize} ref={mesh} />;
 };
