+++ src/r3f/Sun.tsx
@@ -1,24 +1,10 @@
-import React from 'react';
-import { useFrame } from '@react-three/fiber';
-import * as THREE from 'three';
-
 const Sun = () => {
-  const sunRef = React.useRef<THREE.Mesh>(null);
-
-  useFrame(() => {
-    if (sunRef.current) {
-      sunRef.current.rotation.y += 0.01;
-    }
-  });
-
   return (
-    <group>
-      <mesh ref={sunRef}>
-        <sphereGeometry args={[5, 32, 32]} />
-        <meshBasicMaterial color="yellow" emissive="yellow" emissiveIntensity={1} />
-      </mesh>
-      <pointLight color="yellow" intensity={2} distance={50} />
-    </group>
+    <mesh position={[1000, 0, 0]}>
+      <sphereGeometry args={[100, 128, 128]} />
+      <meshStandardMaterial emissive="yellow" emissiveIntensity={10} />
+      <pointLight color="white" intensity={10000000}  />
+    </mesh>
   );
 };
